MDL-71156 cli: Script to report and fix orphaned calendar events
Fix orphaned calendar events that were broken by MDL-67494.

This script will look for all the calendar events which userids
where broken by a wrong upgrade step, affecting to Moodle 3.9.5
and up.

It performs checks to both:
  a) Detect if the site was affected (ran the wrong upgrade step).
  b) Look for orphaned calendar events, categorising them as:
     - standard: site / category / course / group / user events
     - subscription: events created via subscriptions.
     - action: normal action events, created to show common important dates.
     - override: user and group override events, particular, that some activities support.
     - custom: other events, not being any of the above, common or particular.

By specifying it (--fix) try to recover as many broken events (missing userid) as
possible. Standard, subscription, action, override events in core are fully supported but
override or custom events should be fixed by each plugin as far as there isn't any standard
API (plugin-wise) to launch a rebuild of the calendar events.

Historic note:

We have finally removed the facility of searching in logs
for calendar event creation users because it was extremely
slow, with the logging tables not having the needed indexed
(objectid and tablename) needed to run quick.

So we have changed to a simpler approach, as far as the userid
of all those events is just the "creator" of the events, and not
the "target" of the event. It's as simple as this:

- Any site / category event, set userid to admin.
- And course, group, action event, set userid to admin.

Of course, override events (which users are the target)
are apart from this and they are carefully built based on
the information present in the override tables.

20210422 - Amended to match all core changes.