MDL-xxxxx composer: Bump to PHPUnit 10
Only for php81 and up, this is the very first step
towards preparing ourselves for it.

Also, provide autoloading of the new PHPUnit extension
that we'll be using to provide some features previously
implemented via inheritance, now forbidden.

Note that I'm not sure that such an autoloading is the
best solution. I just decided to go that way to be 100%
parallel with the current implementation of the Behat
extension.

At some point we'll have to consider to move these extensions
and may other classes from their current manual behat/phpunit
directories (both core and admin tools), under namespaces.

But that's out from the scope of this issue, so for now, let's
create and autoload the PHPUnit extension similarly to the Behat one.